requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
geoalchemy2==0.14.2
Shapely==2.0.1
python-dotenv==1.0.0
Pillow==10.0.0
Flask-Uploads==0.2.1
python-dateutil==2.8.2